+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## 0.9.1
-- Suppress known-harmless NAT warnings
+- Suppressed known-harmless NAT warnings
 
 ## 0.9.0
 - Removed side-effects in MCPConfig